adios: use adios_db
