+++ environment.yml
@@ -26,6 +26,7 @@ dependencies:
   - nc-time-axis
   - geojson
   - opendrift-landmask-data
+  - pynucos>=2.12
   - oil_library-opendrift>=4+noaa1.1.3
   - isodate
   - coloredlogs
